Make telemetry generation tool more general
