--- tools/perf-test-processing/README.md
@@ -1,37 +1,0 @@
-# @fluid-tools/perf-test-processing
-
-Internal tool to process the output of our performance benchmark tests and write relevant metrics to our monitoring
-systems.
-
-## Running locally
-
-Install dependencies and build the project:
-
-```bash
-npm i
-npm run build
-```
-
-This tool leverages [oclif](https://oclif.io/). To run it manually, you need to call oclif's `run` script in the `bin`
-folder, like so:
-
-```bash
-node bin/run <arguments to the tool>
-```
-
-The tool expects an argument for the type of test that it's going to be processing, and one or more arguments for
-folders that contain the test output files (generated by the `@fluid-tools/benchmark` tool when called with the
-appropriate options to generate output JSON files).
-
-To see all arguments and expected usage run `node bin/run --help`.
-
-The `start` npm script can be used as an alternative to `node bin/run`, with an extra separator for the arguments, like
-so:
-
-```bash
-npm run start -- <arguments to the tool>
-```
-
-Make sure that the files in the folders you specify and the test type match (i.e. run with `--testType executionTime`
-and folders that contain output files for execution-time tests, or `--testType memoryUsage` and folders that contain
-output files for memory usage tests).
