+++ tools/telemetry-generator/.eslintrc.js
