+++ tools/telemetry-generator/LICENSE
