+++ tools/telemetry-generator/README.md
@@ -1,0 +1,51 @@
+# @fluid-tools/telemetry-generator
+
+Internal tool to take JSON files and generate custom telemetry events based on the data in them. Used, for example,
+to process the output of our performance tests and write relevant metrics to our monitoring systems.
+
+## Running the tool
+
+Install dependencies and build the project:
+
+```bash
+npm i
+npm run build
+```
+
+This tool leverages [oclif](https://oclif.io/). To run it manually, you need to call oclif's `run` script in the `bin`
+folder, like so:
+
+```bash
+node bin/run <arguments to the tool>
+```
+
+The `start` npm script can be used as an alternative to `node bin/run`, with an extra separator for the arguments, like
+so:
+
+```bash
+npm run start -- <arguments to the tool>
+```
+
+The tool has two required arguments:
+
+- `--dir`: the path(s) to one or more folders which contain the JSON files to be processed
+- `--handlerModule`: the path to a JavaScript file that exports a handler function which can process those JSON files.
+
+Run `node bin/run --help` for details on the arguments.
+
+## Processing the Fluid Framework performance tests
+
+When this tool is used to process our performance tests, the folders should contain files generated by the
+[`@fluid-tools/benchmark`](../benchmark) tool, and the [handlers defined here](./handlers) can be used to process them.
+Make sure to use the correct type of handler (execution time / memory usage) for the output files you want to process.
+Only one type of file can be processed during a single run of this tool.
+
+## Handling custom files
+
+If you want to process your own JSON files with this tool you just need to pass a custom handler to it. See the
+[handler template](./handlers/template.ts) for details on how that file should look like. Then call the tool as usual,
+e.g.:
+
+```bash
+node bin/run --handlerModule /path/to/your/handler.js --dir /path/to/your/files
+```
