+++ tools/telemetry-generator/bin/dev
