+++ tools/telemetry-generator/bin/dev.cmd
