+++ tools/telemetry-generator/bin/run
