+++ tools/telemetry-generator/bin/run.cmd
