+++ tools/telemetry-generator/src/index.ts
