+++ tools/telemetry-generator/src/logger.ts
