+++ tools/telemetry-generator/tsconfig.json
